♻️ refactor(app): relocate seo utility to lib directory
- move getSEOTags import from utils to lib
- rename seo utility file path for consistency

✨ feat(services): enhance service component styles

- add dynamic border and hover effects to services
- improve layout with responsive design adjustments

✨ feat(utils): add utility function for class name merging

- implement cn function in utils for class name management
- utilize clsx and tailwind-merge for optimized class handling
