+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next"
 import { Geist, Geist_Mono } from "next/font/google"
 
-import { getSEOTags } from "@/utils/seo"
+import { getSEOTags } from "@/lib/seo"
 
 import "@/styles/globals.css"
